Slider fixed and examples/tutorial pages completed up to Conditioanl Logic.

+++ Sources/SwiftProcessing/Core/Sketch3DPrimitives.swift
@@ -14,7 +14,7 @@ public extension Sketch {
         if var shapeNode = self.currentTransformationNode.getAvailableShape(newtag) {
 
         } else {
-            geometry.firstMaterial?.diffuse.contents = UIColor(red: self.settings.fill.red/255.0, green: self.settings.fill.green/255.0, blue: self.settings.fill.blue/255.0, alpha: self.settings.fill.alpha)
+            geometry.firstMaterial?.diffuse.contents = UIColor(red: CGFloat(self.settings.fill.red/255.0), green: CGFloat(self.settings.fill.green/255.0), blue: CGFloat(self.settings.fill.blue/255.0), alpha: CGFloat(self.settings.fill.alpha))
 
             if self.texture != nil && self.textureEnabled {
                 geometry.firstMaterial?.diffuse.contents = self.texture!.currentFrame()
